Weather Data Analysis Cloud Based/cloud_weather_project/requirements.txt
pandas==2.2.2
requests==2.32.3
pyyaml==6.0.2
python-dotenv==1.0.1
pyarrow==17.0.0
boto3==1.34.162
typer==0.12.3
streamlit==1.38.0
plotly==5.24.1
dask[complete]==2024.8.2
